Extract migration batch methods into separate class.
refs #5934
