+++ app/jobs/migration_job.rb
@@ -6,13 +6,13 @@ class MigrationJob < ApplicationJob
 
   # @param migrator_class [Class] the name of the migrator class to run
   # @param batch_descriptor [Array(Integer, Integer)] a [index_or_after_id, limit] pair
-  #   as returned by MigrationRunner.batch_descriptors
+  #   as returned by BatchSupport.batch_descriptors
   # @param mode [Symbol] the mode in which to operate on obj (see README for explanation of modes)
   # @param background_job_result [BackgroundJobResult] an object to store the result of the background job
   def perform(migrator_class:, batch_descriptor:, mode:, background_job_result:)
     background_job_result.processing!
 
-    druids_slice = Migrators::MigrationRunner.druids_for_batch(migrator_class:, batch_descriptor:)
+    druids_slice = Migrators::BatchSupport.druids_for_batch(migrator_class:, batch_descriptor:)
 
     results = Migrators::MigrationRunner.migrate_druid_list(migrator_class:, mode: mode,
                                                             druids_slice:).map do |result|
